Move runtimeVersion to state, add rudimentary test for subscription.
